backend/requirements.txt
pandas
ortools
fastapi
uvicorn
beautifulsoup4
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
gunicorn
